Replace tickersSubscribable with tickersPublisher #4185
